third_party/WebKit/LayoutTests/fast/dom/custom/register-element-with-U212A.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
test(function () {

    var A = document.registerElement('x-tkent');
    assert_true(document.createElement('x-tkent') instanceof A);

    var B = document.registerElement('x-t\u212Aent');
    assert_true(document.createElement('x-t\u212Aent') instanceof B);
	
}, 'registerElement should register type with U+212A');

</script>
